Very first commit (setup & backup purpose). Does not compile yet | TODO:
=> Files to be "Rserve-ified":
* HeatMapTask.java
* CameraSearchTask.java
* CentWaveDetector.java

=> User config:
* Remove R paths env variables from "startMZmineScripts"
* Add the single R executable path to "MZmine config setup dialog"
(Ask for it if and only if the said path could not be detected automatically - fallback case - and if one tried to run a task requiring R)...

Note:
-------
For Rserve cleanup on MZmine or JVM ungraceful exit (crash or kill), or to ensure one won't shutdown other already running Rserve instances (from another app. or from an other MZmine instance), some little changes have been made in the "startMZmineScripts".